Support multiple database backends
Previous to this commit, only the sqlite database backend was
supported.  This commit supports any number of database backends
by utilizing the bacula syntax in the bacula.dir configuration
file.

This commit also supports the management of the mysql database
backend. If manage_db is true and the db_backend is 'mysql'
as well as either is_director or is_storage is true,
the bacula module will create the mysql database, and grant
privileges to the db_user. Further, if manage_db_tables is true
(is true by default) the bacula module will create the database
tables for the db_backend when the bacula server packages are
installed