Add manage_db parameter
Previous to this commit, the db_database had to be created prior
to the creation of the tables.  Using the puppetlabs/mysql module,
the bacula module can now manage the creation of the database for
the user. The feature requires the puppetlabs/mysql module be in
the modulepath on the master as well as the manage_db parameter
be true (defaults to false).